add button / reorg entity inheritance
