+++ hahomematic/const.py
@@ -145,6 +145,7 @@ IGNORED_PARAMETERS = [
     "INCLUSION_UNSUPPORTED_DEVICE",
     "INHIBIT",
     "INSTALL_TEST",
+    "INSTALL_MODE",
     "PARTY_SET_POINT_TEMPERATURE",
     "PARTY_TIME_END",
     "PARTY_TIME_START",
@@ -289,6 +290,7 @@ DEFAULT_VERIFY_TLS = False
 
 HA_PLATFORM_ACTION = "action"
 HA_PLATFORM_BINARY_SENSOR = "binary_sensor"
+HA_PLATFORM_BUTTON = "button"
 HA_PLATFORM_CLIMATE = "climate"
 HA_PLATFORM_COVER = "cover"
 HA_PLATFORM_EVENT = "event"
@@ -302,6 +304,7 @@ HA_PLATFORM_TEXT = "text"
 
 HA_PLATFORMS = [
     HA_PLATFORM_BINARY_SENSOR,
+    HA_PLATFORM_BUTTON,
     HA_PLATFORM_CLIMATE,
     HA_PLATFORM_COVER,
     HA_PLATFORM_LIGHT,
